Enhanced NVMe self-test task: updated APIs for progress tracking, error handling, and result evaluation; refactored data structures for self-test logs and device info.

+++ src/smart_check/common/device_info_parts.rs
@@ -2,7 +2,7 @@ use serde::Deserialize;
 
 #[derive(Debug, Deserialize)]
 pub struct PowerOnTime {
-    pub hours: u32,
+    pub hours: u64,
 }
 
 #[derive(Debug, Deserialize)]
